Renamed folders and namespaces. Added funciotality: sorting by name.
