+++ ToolsManagement/Controllers/MagazineController.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using ToolsManagement.Models.Drill;
+using ToolsManagement.Models.DrillModel;
 using ToolsManagement.Services;
 
 namespace ToolsManagement.Controllers
