+++ ToolsManagement/Models/DrillModel/DrillMagazineQuantityUpdateDto.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ToolsManagement.Models.DrillModel
+{
+    public class DrillMagazineQuantityUpdateDto
+    {
+        public int Quantity { get; set; }
+    }
+}
