+++ ToolsManagement/Models/DrillQuery.cs
@@ -11,5 +11,7 @@ namespace ToolsManagement.Models
         public string SearchPhrase { get; set; }
         public int PageNumber { get; set; }
         public int PageSize { get; set; }
+        public string SortBy { get; set; }
+        public SortDirection SortDirection { get; set; }
     }
 }
