+++ ToolsManagement/Models/Validators/DrillQueryValidator.cs
@@ -4,12 +4,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using ToolsManagement.Data.Entities;
 
 namespace ToolsManagement.Models.Validators
 {
     public class DrillQueryValidator : AbstractValidator<DrillQuery>
     {
         private int[] allowebPageSizes = new[] { 5, 10, 15 };
+        private string[] allowedSortByColumnNames = { nameof(Drill.Name) };
         public DrillQueryValidator()
         {
             RuleFor(r => r.PageNumber).GreaterThanOrEqualTo(1);
@@ -20,6 +22,9 @@ namespace ToolsManagement.Models.Validators
                     context.AddFailure("PageSize", $"PageSize must in [{string.Join(",", allowebPageSizes)}]");
                 }
             });
+
+            RuleFor(r => r.SortBy).Must(value => string.IsNullOrEmpty(value) || allowedSortByColumnNames.Contains(value))
+                .WithMessage($"Sort by is optional or must be in [{string.Join(",", allowedSortByColumnNames)}]");
         }
     }
 }
