+++ ToolsManagement/Services/Interfaces/IDrillService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using ToolsManagement.Entities;
 using ToolsManagement.Models;
-using ToolsManagement.Models.Drill;
+using ToolsManagement.Models.DrillModel;
 
 namespace ToolsManagement.Services.Interfaces
 {
